moving routings from job to data_set in module_utils Minor cleanups Needs re-pipelined
